backend/requirements.txt
# === Core Frameworks ===
fastapi>=0.115.0
uvicorn>=0.30.1
streamlit>=1.38.0

# === LLM + Agents ===
langchain==0.3.4
langchain-core==0.3.13
langchain-community==0.3.3
langchain-groq>=0.1.9
# OpenAI packages removed per project policy (use HuggingFace embeddings and Groq for LLM)

# === Vector Store + Embeddings ===
tiktoken>=0.7.0
chromadb>=0.5.3
sentence-transformers>=2.2.2
tiktoken>=0.7.0

# === PDF Processing ===
PyMuPDF>=1.24.9
pdfminer.six>=20231228

# === Search + Research Tools ===
duckduckgo-search>=6.1.5
serpapi>=0.1.5
arxiv>=2.1.3
datasets>=3.0.1

# === Utilities ===
pydantic>=2.9.2
python-dotenv>=1.0.1
requests>=2.32.3
aiofiles>=24.1.0
numpy>=1.26.4
pandas>=2.2.3


python-multipart
google-search-results